requirements.txt
asgiref==3.5.2
asttokens==2.0.7
backcall==0.2.0
colorama==0.4.5
decorator==5.1.1
Django==4.1
django-admin-interface==0.19.2
django-colorfield==0.7.2
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-model-utils==4.2.0
django-notifications-hq==1.7.0
executing==0.9.1
ipython==8.4.0
jedi==0.18.1
jsonfield==3.1.0
matplotlib-inline==0.1.3
parso==0.8.3
pickleshare==0.7.5
Pillow==9.2.0
prompt-toolkit==3.0.30
pure-eval==0.2.2
Pygments==2.12.0
pytz==2022.2.1
six==1.16.0
sqlparse==0.4.2
stack-data==0.3.0
swapper==1.3.0
traitlets==5.3.0
tzdata==2022.1
wcwidth==0.2.5
